NewEventForm code moved
